test: Add tests for Scanner
Added basic tests for the Scanner class, mocking the ML Kit Scanner API. Slightly refactored Scanner to allow for easier testing.

Integrated the Google ML Kit document scanner API. The Scanner must be initialized in the activity that will call the scan, and then scanning can be launched from a button click or other user interaction. This launches a separate activity, and the resulting scanned document's URI is securely returned to the main activity using FileProvider.

For now, the scanner object is created in MainActivity's OnCreate, and then passed as arguments to the UI elements that call for a scan. If other API integrations require this type of initialization, consider integrating them in a NavigationActions.kt style class for easier initialization and access.