refactor: check for null edited note
Currently, we don't check if the selected note is null when we are editing it or its markdown. Instead, we give default values to the functions that need it, so in case of error (which should not happen) we won't know what's wrong.

Instead of this, I propose opening an error screen if this unexpected should happen, which allows the user to go back, and then if the edit screen is opened we assume (!!) selected note is correct.