fix: remove deprecated GoogleSignIn and refactor sign in
Completely removed the use of the deprecated GoogleSignIn, using Credential Manager instead.
This also involved a lot of refactoring to increase readability of the code, and move most code to the renamed Authenticator file (instead of defining it in Sign In). This now handles sign in with credential manager (using google's federated identity), and authentication with Firebase Auth.

The Authenticator is also moved out of the sign in class to enable correct logout (including the resetting of the credential manager state) in profile screens.

Finally, replaced the creation of unmanaged unconnected coroutines with safer coroutines connected to the composable functions they are called in.