feat: add and fix UI for saved folders
Apart from integrating the UI for saving folders, now public and friends folders only show notes and subfolders that are public or friends. Also fixed tests to conform to the new function signatures that make use of the userViewModel.

To be done still is fix navigation actions that always return to search screen when returning from a folder.